+++ Bin/BuildCodeBlocks.sh
@@ -1,0 +1,7 @@
+cd ..
+if ! [ -e Build ]; then
+    mkdir Build
+fi
+cd Build/
+cmake ../ -G "CodeBlocks - Unix Makefiles"
+cd ..