+++ Source/Jinx.h
@@ -101,7 +101,7 @@ namespace Jinx
 	const uint32_t MajorVersion = 0;
 
 	/// Minor version number
-	const uint32_t MinorVersion = 21;
+	const uint32_t MinorVersion = 22;
 
 	/// Patch number
 	const uint32_t PatchNumber = 0;
@@ -496,6 +496,6 @@ namespace Jinx
 	*/
 	void ShutDown();
 
-};
+}
 
 #endif // JINX_H__